rename parts_per_x -> per_things
